+++ src/pages/knowledgegraph/GraphSearch/export/index.ts
@@ -1,0 +1,1 @@
+export * from './exportImages';
